refactored env and home package location in code
